Update clipboard.md
With reference to ticket https://support.syncfusion.com/support/tickets/390133, I spent a lot of time trying to capture events associated with pasting records into the grid and the conclusion of the ticket was that events are not supported by Syncfusion for the Grid Paste action. I don't think this would be the expected behavior for most users, so believe some comment should be made here.